feat: moving amplience context directory to correct location
  * includes some additional typing issue fixes
